доработаны функции
